+++ vignettes/v1_interactiveplots.R
@@ -136,7 +136,7 @@ static_plot <- static +
     axis.ticks.y     = element_blank(),
     axis.title.x     = element_blank(),
     axis.title.y     = element_blank()
-  ) + scale_color_viridis(
+  ) + scale_color_viridis(option = "inferno",
     discrete = TRUE,
     labels = c("Female", "Male", "Unknown")
   )